refactor: remove league description and unify duplicated league services
