+++ backend/src/tests/integration/league.integration.test.ts
@@ -1,7 +1,11 @@
 import { env } from "cloudflare:workers";
 import { Temporal } from "@js-temporal/polyfill";
 import { describe, it, expect, beforeEach } from "vitest";
-import { LeagueService, GLOBAL_LEAGUE_ID } from "../../services/league";
+import {
+  LeagueService,
+  GLOBAL_LEAGUE_ID,
+  toLeagueDTO,
+} from "../../services/league";
 import { LeagueRepositoryD1 } from "../../repositories/d1/leagueRepositoryD1";
 import { LeagueRepository } from "../../repositories/leagueRepository";
 import { success, failure } from "../../repositories/result";
@@ -22,7 +26,6 @@ describe("LeagueService Integration Tests", () => {
       if (result.ok) {
         expect(result.value.id).toBe(GLOBAL_LEAGUE_ID);
         expect(result.value.title).toBe("Global League");
-        expect(result.value.description).toBeTruthy();
         expect(result.value.domain).toBe("en");
         expect(result.value.icon).toBe("🌍");
         expect(result.value.teams).toEqual([]);
@@ -83,6 +86,34 @@ describe("LeagueService Integration Tests", () => {
           teams: [],
         });
       }
+    });
+  });
+});
+
+describe("toLeagueDTO", () => {
+  it("maps a domain League to the list-endpoint LeagueDTO shape", () => {
+    const startDate = Temporal.Instant.from("2026-01-01T00:00:00Z");
+    const endDate = Temporal.Instant.from("2026-12-31T00:00:00Z");
+    const league: League = {
+      id: "league-7",
+      name: "Trivia Titans",
+      adminId: "player-1",
+      startDate,
+      endDate,
+      domain: "it",
+      icon: "🏆",
+    };
+
+    const dto = toLeagueDTO(league);
+
+    expect(dto).toEqual({
+      id: "league-7",
+      title: "Trivia Titans",
+      domain: "it",
+      icon: "🏆",
+      startDate,
+      endDate,
+      teams: [],
     });
   });
 });
