+++ frontend/src/mocks/data/leagues.ts
@@ -9,8 +9,6 @@ export const leagues: LeagueDTO[] = [
     id: "global",
     title: "Global League",
     icon: "🌍",
-    description:
-      "Compete with players from around the world in the ultimate Wikipedia trading experience!",
     domain: "en",
     startDate: Instant.from("2024-01-01T00:00:00Z"),
     endDate: Instant.from("2024-12-31T23:59:59Z"),
@@ -20,7 +18,6 @@ export const leagues: LeagueDTO[] = [
     id: "italy",
     title: "Italia League",
     icon: "🍕",
-    description: "La lega italiana",
     domain: "it",
     startDate: Instant.from("2024-01-01T00:00:00Z"),
     endDate: Instant.from("2024-02-28T23:59:59Z"),
@@ -30,7 +27,6 @@ export const leagues: LeagueDTO[] = [
     id: "europe",
     title: "Europe League",
     icon: "🇪🇺",
-    description: "Compete across Europe",
     domain: "en",
     startDate: Instant.from("2024-01-01T00:00:00Z"),
     endDate: Instant.from("2024-03-15T23:59:59Z"),
@@ -40,7 +36,6 @@ export const leagues: LeagueDTO[] = [
     id: "americas",
     title: "Americas League",
     icon: "🌎",
-    description: "The Americas league",
     domain: "en",
     startDate: Instant.from("2024-01-01T00:00:00Z"),
     endDate: Instant.from("2024-03-20T23:59:59Z"),
